pyproject.toml
[project]
name = "classic-cars-pydanticai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "gradio>=5.27.0",
    "pydantic-ai-slim[mcp,openai]>=0.1.4",
]
